Requests metadata about Cromwell workflow jobs during a time period specified

Requests metadata about Cromwell workflow jobs during a time period specified

Usage

cromwell_jobs(
  days = 1,
  workflow_name = NULL,
  workflow_status = NULL,
  url = cw_url(),
  token = NULL
)

Arguments

days

The number of days of history to return, defaults to 1 day.

workflow_name

An array of strings of valid workflow names you want in your job list.

workflow_status

A array of strings of valid workflow statuses you want in your job list (e.g., submitted, running, succeeded, failed, aborting, aborted)

url

(character) base url for your Cromwell server. optional. if not supplied set the url as the env var CROMWELLURL

token

(character) we do not recommend passing your token here as a string. Either pass it using Sys.getenv() or save your token as the env var PROOF_TOKEN and then passing nothing to this param and we'll find it

Value

a tibble of metadata on workflow jobs submitted; each row is a different job. columns:

end (dttm)

workflow_id (chr)

metadataArchiveStatus (chr)

workflow_name (chr)

start (dttm)

status (chr)

submission (dttm)

workflowDuration (dbl)

Important

Requires valid Cromwell server URL to be set. See cromwell_settings

Author

Amy Paguirigan, Scott Chamberlain

Examples

if (FALSE) { # \dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
} # }