Cromwell Settings

Cromwell Settings

Cromwell URL

A URL for the Cromwell server you're interacting with is required. The only way to do so is by setting the env var:

CROMWELLURL

You can set this env var in many different ways. From within R you can do this with cromwell_config() like:

cromwell_config(url = "your/url")

Which is equivalent to

Sys.setenv(CROMWELLURL = "your/url")

Note that this only sets the url for the current R session and does not persist among R sessions.

Other options include:

Set CROMWELLURL env var in a .bash_profile file or equivalent that will be picked up by R when it starts.

Set CROMWELLURL env var in an IDE such as RStudio.

Set CROMWELLURL env var by prefixing R on the command line, e.g., CROMWELLURL=http://your/url R

Verbose messaging

You can set your preference for whether you get messages informing you of various tasks rcromwell is performing. See cromwell_config()