Submit a workflow job to Cromwell

Supports the submission of a fully defined workflow job to a Cromwell instance.

Usage

cromwell_submit_batch(
  wdl,
  batch = NULL,
  params = NULL,
  options = NULL,
  labels = NULL,
  dependencies = NULL,
  url = cw_url(),
  token = NULL
)

Arguments

wdl

Local path to the wdl file describing the workflow. (Required)

batch

Local path to the json containing a reference to any batch file desired if the workflow is a batch. (Optional)

params

Local path to the json containing the parameters to use with the workflow. (Optional)

options

Local path to the json containing workflow options to apply. (Optional)

labels

A data frame containing the labels for this workflow. (Optional)

dependencies

A zip'd file of subworkflow dependencies. (Optional)

url

(character) base url for your Cromwell server. optional. if not supplied set the url as the env var CROMWELLURL

token

(character) we do not recommend passing your token here as a string. Either pass it using Sys.getenv() or save your token as the env var PROOF_TOKEN and then passing nothing to this param and we'll find it

Value

a tibble with one row and two columns:

id: workflow ID you'll need to monitor the job

status: the status of the job

Important

Requires valid Cromwell server URL to be set. See cromwell_settings

Author

Amy Paguirigan, Scott Chamberlain